# Conflicts:
#	cmd/gen-commands.ts
#	docs/md/Version-history.md
#	src/Accelerators.cpp
#	src/Canvas.cpp
#	src/Commands.cpp
#	src/Commands.h
#	src/RefHoverDetect.cpp
#	src/RefHoverDetect.h
#	src/WindowTab.cpp
#	src/WindowTab.h

+++ src/base/ByteReaderWriter.cpp
@@ -303,6 +303,8 @@ bool ByteReader::UnpackBE(void* strct, int size, Str format, int off) const {
     return Unpack(strct, size, format, off, true);
 }
 
+// Unpacks a structure from the data according to the given format
+// e.g. the format "32b2w6d" unpacks 32 Bytes, 2 16-bit Words and 6 32-bit Dwords
 bool ByteReader::Unpack(void* strct, int size, Str format, bool isBE, int off) const {
     return Unpack(strct, size, format, off, isBE);
 }
@@ -326,7 +328,7 @@ u32 UInt32LE(const u8* d) {
 // --- ByteWriter
 
 ByteWriter::ByteWriter(int sizeHint) {
-    d.cap = (u32)sizeHint;
+    d.cap = sizeHint;
 }
 
 void ByteWriter::Write8(u8 b) {
@@ -383,6 +385,6 @@ Str ByteWriter::AsByteSlice() const {
 }
 
 ByteWriterLE::ByteWriterLE(int sizeHint) {
-    d.cap = (u32)sizeHint;
+    d.cap = sizeHint;
     isLE = true;
 }
